unit test for project handlers complete removed all mixture of unit/integration tests
